# docs/swagger.yaml
definitions:
...
  models.CalculateExerciseGrades:
    properties:
      exercise_id:
        type: integer
      student_id:
        type: integer
    type: object
...
  models.CreateExamsRequest:
    properties:
      class_id:
        type: integer
      content: {}
      end_time:
        type: string
      start_time:
        type: string
      teacher_id:
        type: integer
      title:
        type: string
      total_marks:
        type: integer
    type: object
  models.CreateExerciseAnswersRequest:
    properties:
      answers: {}
      exercise_id:
        type: integer
      student_id:
        type: integer
    type: object
...
  models.Exams:
    properties:
      class_id:
        type: integer
      content: {}
      end_time:
        type: string
      id:
        type: integer
      start_time:
        type: string
      status:
        type: string
      teacher_id:
        type: integer
      title:
        type: string
      total_marks:
        type: integer
    type: object
  models.ExerciseAnswers:
    properties:
      answers: {}
      exercise_id:
        type: integer
      id:
        type: integer
      status:
        type: string
      student_id:
        type: integer
    type: object
  models.ExerciseGrades:
    properties:
      detail: {}
      exercise_id:
        type: integer
      id:
        type: integer
      score:
        type: number
      student_id:
        type: integer
    type: object
...
info:
  contact: {}
  description: API documentation for Project PPL - Learnify
  title: Project PPL API
  version: 0.2.1
paths:
...
  /api/v1/exams:
    delete:
      consumes:
      - application/json
      description: Delete an exam by ID
      parameters:
      - description: Exam ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Exam
      tags:
      - Exams
    get:
      consumes:
      - application/json
      description: Fetch exam from a class
      parameters:
      - description: Class ID
        in: query
        name: class_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Exams'
            type: array
      security:
      - BearerAuth: []
      summary: Get Exams by Class ID
      tags:
      - Exams
    patch:
      consumes:
      - application/json
      description: Update an existing exam
      parameters:
      - description: Exam ID
        in: query
        name: id
        required: true
        type: integer
      - description: Updated data
        in: body
        name: exam
        required: true
        schema:
          $ref: '#/definitions/models.CreateExamsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Exams'
      security:
      - BearerAuth: []
      summary: Update Exam
      tags:
      - Exams
    post:
      consumes:
      - application/json
      description: Create a new exam
      parameters:
      - description: Exam data
        in: body
        name: exam
        required: true
        schema:
          $ref: '#/definitions/models.CreateExamsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Exams'
      security:
      - BearerAuth: []
      summary: Create Exam
      tags:
      - Exams
...
  /api/v1/exams/student:
    get:
      consumes:
      - application/json
      description: Fetch all exams for a specific class
      parameters:
      - description: Class ID
        in: query
        name: class_id
        required: true
        type: integer
      - description: Number
        in: query
        name: number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Exams'
            type: array
      security:
      - BearerAuth: []
      summary: Get Exams for Student
      tags:
      - Exams
...
  /api/v1/exercises-answers:
    delete:
      consumes:
      - application/json
      description: Delete an exercise answer by ID
      parameters:
      - description: Exercise ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Exercise Answer
      tags:
      - Exercise Answers (Student Answers)
    get:
      consumes:
      - application/json
      description: Fetch all exercises for a specific material
      parameters:
      - description: Exercise ID
        in: query
        name: exercise_id
        required: true
        type: integer
      - description: Student ID
        in: query
        name: student_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ExerciseAnswers'
            type: array
      security:
      - BearerAuth: []
      summary: Get Exercise Answers by Exercise ID
      tags:
      - Exercise Answers (Student Answers)
    patch:
      consumes:
      - application/json
      description: Update an existing exercise answer
      parameters:
      - description: Exercise Answer ID (Student's Answer ID)
        in: query
        name: id
        required: true
        type: integer
      - description: Updated data
        in: body
        name: exercise
        required: true
        schema:
          $ref: '#/definitions/models.CreateExerciseAnswersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ExerciseAnswers'
      security:
      - BearerAuth: []
      summary: Update Exercise Answer
      tags:
      - Exercise Answers (Student Answers)
    post:
      consumes:
      - application/json
      description: Create a new exercise
      parameters:
      - description: Exercise answers data
        in: body
        name: exercise
        required: true
        schema:
          $ref: '#/definitions/models.CreateExerciseAnswersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ExerciseAnswers'
      security:
      - BearerAuth: []
      summary: Create Exercise Answers
      tags:
      - Exercise Answers (Student Answers)
  /api/v1/exercises/calculate-grade:
    post:
      consumes:
      - application/json
      description: Calculate grade after submitting exercise
      parameters:
      - description: Exercise grades data
        in: body
        name: exercise
        required: true
        schema:
          $ref: '#/definitions/models.CalculateExerciseGrades'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ExerciseGrades'
      security:
      - BearerAuth: []
      summary: Calculate Grade
      tags:
      - Exercise Answers (Student Answers)
  /api/v1/exercises/get-all-grade:
    get:
      consumes:
      - application/json
      description: Get calculated grades for all materials
      parameters:
      - description: Student ID
        in: query
        name: student_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ExerciseGrades'
      security:
      - BearerAuth: []
      summary: Get All Grades
      tags:
      - Exercise Answers (Student Answers)
  /api/v1/exercises/get-grade:
    get:
      consumes:
      - application/json
      description: Get calculated grade
      parameters:
      - description: Exercise ID
        in: query
        name: exercise_id
        required: true
        type: integer
      - description: Student ID
        in: query
        name: student_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ExerciseGrades'
      security:
      - BearerAuth: []
      summary: Get Grade
      tags:
      - Exercise Answers (Student Answers)
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}" to authenticate
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"